corrected the launch of the app and the login page
made sure that the login page was pulling from the correct contexts, uses the auth properly (need to implement the database usage in the code after its set).
changed some styles for clear visibillity until we lock onto what theme we're going for.
also, changed the configuration of the vite-related part of the app, so there is a need to nmp install again